[splines-major-2]
* `SPLINE_.py`
    - removed `x_values` and `get_splines` from `SPLINE` class.
    - moved `QuadraticSpline` and `NaturalCubicSpline`.
* `SplineInterpolation.py`
    - cleaned up the script
    - added functions for linear, quadratic and natural cubic spline
