+++ src/Solvers/interpolators/__init__.py
@@ -1,6 +1,1 @@
 """Created on Dec 20 13:36:09 2023"""
-
-from typing import List
-
-L_FLOAT = List[float]
-TOLERANCE = 1e-10
