+++ src/Solvers/interpolators/_backend/__init__.py
@@ -1,1 +1,3 @@
 """Created on Dec 20 13:35:59 2023"""
+
+TOLERANCE = 1e-10
